+++ src/components/TeamGettingPage.tsx
@@ -2,8 +2,7 @@ import * as React from "react";
 
 import Spinner from "react-bootstrap/Spinner";
 
-import { DownloadPage } from "./DownloadPageV2";
-import { Instructions } from "./Instructions";
+import { DownloadPage } from "./DownloadPage";
 
 import { getCurrentTeams, TeamInfo } from "../services/CurrentTeams";
 import { useEffect, useState } from "react";
@@ -18,7 +17,6 @@ export const TeamGettingPage: React.FunctionComponent = () => {
             {teams ? (
                 <>
                     <DownloadPage teamInfo={teams} />
-                    <Instructions />
                 </>
             ) : (
                 <div style={{ width: "100%", display: "flex", justifyContent: "center", marginTop: "1em" }}>
